Move tikTok state to common
