+++ src/containers/upload-page/components/ShareBanner.tsx
@@ -9,10 +9,10 @@ import { ReactComponent as IconShare } from 'assets/img/iconShare.svg'
 import { useModalState } from 'common/hooks/useModalState'
 import { Name } from 'common/models/Analytics'
 import User from 'common/models/User'
+import { open as openTikTokModal } from 'common/store/ui/share-sound-to-tiktok-modal/slice'
 import Toast from 'components/toast/Toast'
 import { MountPlacement, ComponentPlacement } from 'components/types'
 import { useFlag } from 'containers/remote-config/hooks'
-import { open as openTikTokModal } from 'containers/share-sound-to-tiktok-modal/store/slice'
 import AudiusBackend from 'services/AudiusBackend'
 import apiClient from 'services/audius-api-client/AudiusAPIClient'
 import { FeatureFlags } from 'services/remote-config'
