+++ src/store/types.ts
@@ -2,6 +2,7 @@ import { RouterState } from 'connected-react-router'
 
 import { CommonState } from 'common/store'
 import modals from 'common/store/ui/modals/slice'
+import { ShareSoundToTikTokModalState } from 'common/store/ui/share-sound-to-tiktok-modal/types'
 import { AddToPlaylistState } from 'containers/add-to-playlist/store/reducers'
 import ArtistDashboardState from 'containers/artist-dashboard-page/store/types'
 import ArtistRecommendationsReducer from 'containers/artist-recommendations/store/slice'
@@ -34,7 +35,6 @@ import SearchBarState from 'containers/search-bar/store/types'
 import SearchPageState from 'containers/search-page/store/types'
 import ServiceSelectionReducer from 'containers/service-selection/store/slice'
 import SettingsPageState from 'containers/settings-page/store/types'
-import { ShareSoundToTikTokModalState } from 'containers/share-sound-to-tiktok-modal/store/types'
 import SignOnPageState from 'containers/sign-on/store/types'
 import { SmartCollectionState } from 'containers/smart-collection/store/slice'
 import TrackPageState from 'containers/track-page/store/types'
